feat: Can a los botones de las vistas

+++ resources/views/livewire/admin/coursedetails/index.blade.php
@@ -14,14 +14,15 @@
 
     <div class="space-y-2">
 
-        <!-- Botón de nuevo -->
-        <div>
-            <x-jet-secondary-button wire:click="create()"
-                class="border-[#1b396a] text-sky-700 hover:text-sky-500 active:text-[#1b396a] active:bg-sky-50">
-                <x-icon.plus solid alt="sm" class="inline-block h-5 w-5" />
-                Aregar detalles de curso
-            </x-jet-secondary-button>
-        </div>
+        @can('coursedetails.create')
+            <div>
+                <x-jet-secondary-button wire:click="create()"
+                    class="border-[#1b396a] text-sky-700 hover:text-sky-500 active:text-[#1b396a] active:bg-sky-50">
+                    <x-icon.plus solid alt="sm" class="inline-block h-5 w-5" />
+                    Aregar detalles de curso
+                </x-jet-secondary-button>
+            </div>
+        @endcan
 
         <div class="flex flex-col sm:flex-row sm:items-baseline sm:gap-x-1.5 pb-6">
             <div class="mt-4 flex-1">
@@ -88,17 +89,23 @@
                         <x-table.cell>{{ $d->lugar }}</x-table.cell>
                         <x-table.cell>{{ $d->grupo }}</x-table.cell>
                         <x-table.cell width='200' class="whitespace-nowrap">
-                            <button  wire:click="view({{ $d->id }})" type="button" title="Ver más información" class="px-4 bg-white hover:text-white hover:bg-[#1b396a] text-black font-bold border border-sky-400 rounded shadow" >
-                                Ver
-                            </button>
-                            <button  wire:click="edit({{ $d->id }})" type="button" title="Editar información" class="mx-2 px-4 bg-white hover:text-white hover:bg-amber-500 text-black font-bold border border-amber-400 rounded shadow" >
-                                Editar
-                            </button>
+                            @can('coursedetails.show')
+                                <button  wire:click="view({{ $d->id }})" type="button" title="Ver más información" class="px-4 bg-white hover:text-white hover:bg-[#1b396a] text-black font-bold border border-sky-400 rounded shadow" >
+                                    Ver
+                                </button>
+                            @endcan
+                            @can('coursedetails.edit')
+                                <button  wire:click="edit({{ $d->id }})" type="button" title="Editar información" class="mx-2 px-4 bg-white hover:text-white hover:bg-amber-500 text-black font-bold border border-amber-400 rounded shadow" >
+                                    Editar
+                                </button>
+                            @endcan
                             {{$this->permiso_para_eliminar($d->id)}}
                             @if($this->permiso_eliminacion)
-                            <button wire:click="delete_details('{{ $d->id }}')" type="button" title="Eliminar detalles" class="px-4 bg-white hover:text-white hover:bg-red-600 text-black font-bold border border-red-400 rounded shadow">
-                                Eliminar
-                            </button>
+                                @can('coursedetails.delete')
+                                    <button wire:click="delete_details('{{ $d->id }}')" type="button" title="Eliminar detalles" class="px-4 bg-white hover:text-white hover:bg-red-600 text-black font-bold border border-red-400 rounded shadow">
+                                        Eliminar
+                                    </button>
+                                @endcan  
                             @endif
                         </x-table.cell>
                     </tr>
